requirements.txt
Django==3.1.7
django-debug-toolbar==3.2
Pillow==8.1.0
requests==2.25.1
WeasyPrint==52.2

django-filter~=2.4.0